testOne/requirements.txt
amqp==5.1.1
asgiref==3.5.2
autopep8==1.7.0
billiard==3.6.4.0
boto3==1.24.68
botocore==1.27.68
cachetools==5.2.0
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
colored==1.4.2
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.4
Django==3.2.13
django-celery-beat==2.3.0
django-cors-headers==3.13.0
django-extensions==3.2.0
django-import-export==2.8.0
django-rest-passwordreset==1.2.1
django-ses==3.1.2
django-sesame==3.1
django-timezone-field==5.0
djangorestframework==3.13.1
et-xmlfile==1.1.0
filelock==3.7.1
google-api-core==2.8.2
google-api-python-client==2.54.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4
httplib2==0.20.4
idna==3.3
importlib-metadata==4.12.0
jmespath==1.0.1
kombu==5.2.4
MarkupPy==1.14
mysqlclient==2.1.1
numpy==1.23.1
oauthlib==3.2.0
odfpy==1.4.1
openpyxl==3.0.10
pandas==1.4.3
platformdirs==2.5.2
prompt-toolkit==3.0.30
protobuf==4.21.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pymorse==1.4
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
toml==0.10.2
tqdm==4.48.2
typing_extensions==4.2.0
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.10
vine==5.0.0
virtualenv==20.15.1
wcwidth==0.2.5
xlrd==2.0.1
xlwt==1.3.0
zipp==3.8.1
redis==4.3.4
